+++ server/resolvers-custom.js
@@ -2,11 +2,13 @@
 
 import fs from "fs";
 import { Kafka, Producer } from "kafkajs";
+import tracing, { Span } from "opentracing";
 import Arango from "./arango";
 import type { QConfig } from "./config";
 import { ensureProtocol } from "./config";
 import path from 'path';
 import fetch from 'node-fetch';
+import { QTracer } from "./tracer";
 
 function isObject(test: any): boolean {
     return typeof test === 'object' && test !== null;
@@ -34,6 +36,7 @@ type Request = {
 type Context = {
     db: Arango,
     config: QConfig,
+    remoteAddress?: string,
     shared: Map<string, any>,
 }
 
@@ -47,69 +50,50 @@ function info(): Info {
 }
 
 async function getAccountsCount(_parent, args, context: Context): Promise<number> {
-    const span = await context.db.tracer.startSpanLog(
-        context,
-        "resolvers-custom.js:getAccountCount",
-        "new getAccountCount query",
-        args);
-    try {
-        const result: any = await context.db.fetchQuery(`RETURN LENGTH(accounts)`, {}, span);
+    const tracer = context.db.tracer;
+    return QTracer.trace(tracer, 'getAccountsCount', async () => {
+        const result: any = await context.db.query(`RETURN LENGTH(accounts)`, {});
         const counts = (result: number[]);
         return counts.length > 0 ? counts[0] : 0;
-    } finally {
-        await span.finish();
-    }
+    }, QTracer.getParentSpan(tracer, context))
 }
 
 async function getTransactionsCount(_parent, args, context: Context): Promise<number> {
-    const span = await context.db.tracer.startSpanLog(
-        context,
-        "resolvers-custom.js:getTransactionsCount",
-        "new getTransactionsCount query",
-        args);
-    try {
-        const result: any = await context.db.fetchQuery(`RETURN LENGTH(transactions)`, {}, span);
+    const tracer = context.db.tracer;
+    return QTracer.trace(tracer, 'getTransactionsCount', async () => {
+        const result: any = await context.db.query(`RETURN LENGTH(transactions)`, {});
         const counts = (result: number[]);
         return counts.length > 0 ? counts[0] : 0;
-    } finally {
-        await span.finish();
-    }
+    }, QTracer.getParentSpan(tracer, context))
 }
 
 async function getAccountsTotalBalance(_parent, args, context: Context): Promise<String> {
-    /*
-    Because arango can not sum BigInts we need to sum separately:
-    hs = SUM of high bits (from 24-bit and higher)
-    ls = SUM of lower 24 bits
-    And the total result is (hs << 24) + ls
-     */
-    const span = await context.db.tracer.startSpanLog(
-        context,
-        "resolvers-custom.js:getAccountTotalBalance",
-        "new getAccountTotalBalance query",
-        args);
-    try {
-        const result: any = await context.db.fetchQuery(`
-        LET d = 16777216
-        FOR a in accounts
-        LET b = TO_NUMBER(CONCAT("0x", SUBSTRING(a.balance, 2)))
-        COLLECT AGGREGATE
-            hs = SUM(FLOOR(b / d)),
-            ls = SUM(b % (d - 1))
-        RETURN { hs, ls }
-    `, {}, span);
+    const tracer = context.db.tracer;
+    return QTracer.trace(tracer, 'getAccountsTotalBalance', async () => {
+        /*
+        Because arango can not sum BigInt's we need to sum separately:
+        hs = SUM of high bits (from 24-bit and higher)
+        ls = SUM of lower 24 bits
+        And the total result is (hs << 24) + ls
+         */
+        const result: any = await context.db.query(`
+            LET d = 16777216
+            FOR a in accounts
+            LET b = TO_NUMBER(CONCAT("0x", SUBSTRING(a.balance, 2)))
+            COLLECT AGGREGATE
+                hs = SUM(FLOOR(b / d)),
+                ls = SUM(b % (d - 1))
+            RETURN { hs, ls }
+        `, {});
         const parts = (result: { hs: number, ls: number }[])[0];
         //$FlowFixMe
         return (BigInt(parts.hs) * BigInt(0x1000000) + BigInt(parts.ls)).toString();
-    } finally {
-        await span.finish();
-    }
+    }, QTracer.getParentSpan(tracer, context))
 }
 
 // Mutation
 
-async function postRequestsUsingRest(requests: Request[], context: Context, rootSpan: any): Promise<void> {
-    const span = await context.db.tracer.startSpan(rootSpan, "resolvers-custom.js:postRequestsUsingRest");
+async function postRequestsUsingRest(requests: Request[], context: Context, span: Span): Promise<void> {
     const config = context.config.requests;
     const url = `${ensureProtocol(config.server, 'http')}/topics/${config.topic}`;
     const response = await fetch(url, {
@@ -131,18 +115,11 @@ async function postRequestsUsingRest(requests: Request[], context: Context, root
     });
     if (response.status !== 200) {
         const message = `Post requests failed: ${await response.text()}`;
-        await span.log({
-            event: 'post request to rest failed ',
-            value: message
-        });
-        await span.finish();
         throw new Error(message);
     }
-    await span.finish();
 }
 
-async function postRequestsUsingKafka(requests: Request[], context: Context, rootSpan: any): Promise<void> {
-    const span = await context.db.tracer.startSpan(rootSpan, "resolvers-custom.js:postRequestUsingKafka");
+async function postRequestsUsingKafka(requests: Request[], context: Context, span: Span): Promise<void> {
     const ensureShared = async (name, createValue: () => Promise<any>) => {
         if (context.shared.has(name)) {
             return context.shared.get(name);
@@ -163,54 +140,43 @@ async function postRequestsUsingKafka(requests: Request[], context: Context, roo
         return newProducer;
 
     });
-    console.log('[postRequests]', requests);
-    await span.log({
-        event: 'post requests to kafka',
-        value: requests
+    const messages = requests.map((request) => {
+        const keyBuffer = Buffer.from(request.id, 'base64');
+        const traceBuffer = Buffer.from([]);
+        context.db.tracer.inject(span, tracing.FORMAT_BINARY, traceBuffer);
+        return {
+            key: Buffer.concat([keyBuffer, traceBuffer]),
+            value: Buffer.from(request.body, 'base64'),
+        };
     });
-    const messages = requests.map((request) => ({
-        key: Buffer.from(request.id, 'base64'),
-        value: Buffer.from(request.body, 'base64'),
-    }));
     await producer.send({
         topic: config.topic,
         messages,
     });
-    await span.log({
-        event: 'messages sended to kafka',
-        value: messages
-    });
-    await span.finish();
 }
 
 async function postRequests(_, args: { requests: Request[] }, context: Context): Promise<string[]> {
-    const span = await context.db.tracer.startSpanLog(
-        context,
-        "resolvers-custom.js:postRequests",
-        "new post request",
-        args);
     const requests: ?(Request[]) = args.requests;
     if (!requests) {
-        await span.finish();
         return [];
     }
-    try {
-        if (context.config.requests.mode === 'rest') {
-            await postRequestsUsingRest(requests, context, span);
-        } else {
-            await postRequestsUsingKafka(requests, context, span);
+
+    const tracer = context.db.tracer;
+    return QTracer.trace(tracer, "postRequests", async (span: Span) => {
+        span.setTag('params', requests);
+        try {
+            if (context.config.requests.mode === 'rest') {
+                await postRequestsUsingRest(requests, context, span);
+            } else {
+                await postRequestsUsingKafka(requests, context, span);
+            }
+            context.db.log.debug('postRequests', 'POSTED', args, context.remoteAddress);
+        } catch (error) {
+            context.db.log.debug('postRequests', 'FAILED', args, context.remoteAddress);
+            throw error;
         }
-    } catch (error) {
-        console.log('[Q Server] post request failed]', error);
-        await span.log({
-            event: 'post request failed',
-            value: error
-        });
-        await span.finish();
-        throw error;
-    }
-    await span.finish();
-    return requests.map(x => x.id);
+        return requests.map(x => x.id);
+    }, QTracer.getParentSpan(tracer, context));
 }
 
 const resolversCustom = {