+++ server/resolvers-mam.js
@@ -1,11 +1,10 @@
 // @flow
 
 import fs from "fs";
-import Arango from "./arango";
-import type { FieldSelection } from "./arango-collection";
-import { Collection, CollectionListener, SubscriptionListener } from "./arango-collection";
-import type { QConfig } from "./config";
 import path from 'path';
+import Arango from "./arango";
+import { Collection, CollectionListener, selectionToString, SubscriptionListener } from "./arango-collection";
+import type { QConfig } from "./config";
 
 type Context = {
     db: Arango,
@@ -53,15 +52,6 @@ function info(): Info {
     };
 }
 
-function selectionToString(selection: FieldSelection[]): string {
-    return selection
-        .filter(x => x.name !== '__typename')
-        .map((field: FieldSelection) => {
-            const fieldSelection = selectionToString(field.selection);
-            return `${field.name}${fieldSelection !== '' ? ` { ${fieldSelection} }` : ''}`;
-        }).join(' ');
-}
-
 function stat(_parent: any, _args: any, context: Context): Stat {
     const listenerToStat = (listener: CollectionListener ): ListenerStat => {
         return {
@@ -93,13 +83,7 @@ function stat(_parent: any, _args: any, context: Context): Stat {
     };
 }
 
-function getChangeLog(_parent: any, args: { id: string }, context: Context): number[] {
-    return context.db.changeLog.get(args.id);
-}
-
 async function getCollections(_parent: any, _args: any, context: Context): Promise<CollectionSummary[]> {
-    const span = await context.db.tracer.startSpanLog(context, "resolvers-mam.js:getCollections",
-        'new getCollections query', _args);
     const db: Arango = context.db;
     const collections: CollectionSummary[] = [];
     for (const collection of db.collections) {
@@ -114,31 +98,17 @@ async function getCollections(_parent: any, _args: any, context: Context): Promi
             indexes,
         });
     }
-    await span.finish();
     return collections;
 }
 
 // Mutation
 
-function setChangeLog(_parent: any, args: { op: string }, context: Context): number {
-    if (args.op === 'CLEAR') {
-        context.db.changeLog.clear();
-    } else if (args.op === 'ON') {
-        context.db.changeLog.enabled = true;
-    } else if (args.op === 'OFF') {
-        context.db.changeLog.enabled = false;
-    }
-    return 1;
-}
-
 export const resolversMam = {
     Query: {
         info,
-        getChangeLog,
         getCollections,
         stat
     },
     Mutation: {
-        setChangeLog,
     },
 };